+++ Sources/StreamChat/Database/DatabaseSession.swift
@@ -81,9 +81,16 @@ protocol CurrentUserDatabaseSession {
     /// Saves the push preference for the given id.
     /// - Parameters:
     ///   - id: The channel ID or the currentUser ID.
-    ///   - payload: The push preference payload.
+    ///   - payload: The push preference.
     @discardableResult
-    func savePushPreference(id: String, payload: PushPreferencePayload) throws -> PushPreferenceDTO
+    func savePushPreference(id: String, payload: PushPreference) throws -> PushPreferenceDTO
+
+    /// Saves the chat preferences for the given id.
+    /// - Parameters:
+    ///   - id: The channel ID or the currentUser ID.
+    ///   - payload: The chat preferences.
+    @discardableResult
+    func saveChatPreferences(id: String, payload: ChatPreferences) throws -> ChatPreferencesDTO
 
     /// Removes the device with the given id from DB.
     func deleteDevice(id: DeviceId)